+++ libAudio/loadOptimFROG.cpp
@@ -192,7 +192,7 @@ bool optimFROG_t::isOptimFROG(const int32_t fd) noexcept
 	std::array<char, 4> optimFrogMagic{};
 	return
 		fd != -1 &&
-		read(fd, optimFrogMagic.data(), optimFrogMagic.size()) == optimFrogMagic.size() &&
+		static_cast<size_t>(read(fd, optimFrogMagic.data(), optimFrogMagic.size())) == optimFrogMagic.size() &&
 		lseek(fd, 0, SEEK_SET) == 0 &&
 		optimFrogMagic == libAudio::optimFROG::magic;
 }
