+++ libAudio/loadWavPack.cpp
@@ -310,7 +310,7 @@ bool wavPack_t::isWavPack(const int32_t fd) noexcept
 	std::array<char, 4> wavPackMagic;
 	return
 		fd != -1 &&
-		read(fd, wavPackMagic.data(), wavPackMagic.size()) == wavPackMagic.size() &&
+		static_cast<size_t>(read(fd, wavPackMagic.data(), wavPackMagic.size())) == wavPackMagic.size() &&
 		lseek(fd, 0, SEEK_SET) == 0 &&
 		wavPackMagic == libAudio::wavPack::magic;
 }
